# Workaround for MAGMA accessing illegal memory in batched cholesky