@mruberry 2021-01-25
This looks great, nice work @heitorschueroff! cc @ptrblck and @xwang233.

I made one comment suggestion and have a question about a change of transpose variant.