@ptrblck 2021-01-25
Thanks for the PR and the discussion on Slack!

Were you somehow able to verify, that the data/results won't be corrupted by this change?
I might be too paranoid, but am afraid of silent errors, which might create wrong results, if we are not only reading out of bounds, but also process these values.

I've spent some time digging into the code and think the first issue is this illegal access:
```
__global__ void spotf2_smlpin_fixwidth_kernel_batched(int m,
        float **dA_array, int ai, int aj, int lda,
        int localstep, int gbstep, magma_int_t *info_array, const int batchCount)
{
    const int batchid = blockIdx.z * blockDim.y + threadIdx.y;
    float *dA = dA_array[batchid] + aj * lda + ai;
    if (batchid >= batchCount) return;
```
`dA_array` has a valid length of 4, while `batchid` is checked too late, so I think swapping it should solve this first issue.
```
(cuda-gdb) print dA_array[0]
$36 = (@generic float * @generic) 0x7ffeef5ff9c0
(cuda-gdb) print dA_array[1]
$37 = (@generic float * @generic) 0x7ffeef5ffb50
(cuda-gdb) print dA_array[2]
$38 = (@generic float * @generic) 0x7ffeef5ffce0
(cuda-gdb) print dA_array[3]
$39 = (@generic float * @generic) 0x7ffeef5ffe70
(cuda-gdb) print dA_array[4]
$40 = (@generic float * @generic) 0xfffffff389
(cuda-gdb) print dA_array[0] @ 4
$44 = {0x7ffeef5ff9c0, 0x7ffeef5ffb50, 0x7ffeef5ffce0, 0x7ffeef5ffe70}
(cuda-gdb) print dA_array[0] @ 5
Error: Failed to read generic memory at address 0x7ffeef9fffe0 on device 0 sm 0 warp 3 lane 0, error=CUDBG_ERROR_MEMORY_MAPPING_FAILED(0x9).
```

However, once this is fixed, another IMA is triggered:
```
(cuda-gdb) bt
#0  0x00005555a86e9a20 in _INTERNAL_61_tmpxft_00007a08_00000000_11_spotf2_kernels_compute_80_cpp1_ii_7b34eafc::sgemm_v20_1_anywidth_device (m=2, n=2,
    k=8, A0=0xff00000000, lda=10, sC=0x7fff260027d0, sB=0x7fffffffff00)
    at /opt/conda/conda-bld/magma-cuda110_1611350834581/work/magmablas/spotf2_devicesfunc.cuh:323
#1  0x00005555a86e9a20 in _INTERNAL_61_tmpxft_00007a08_00000000_11_spotf2_kernels_compute_80_cpp1_ii_7b34eafc::sgemm_v20_1_anywidth_device (m=8, n=0,
    k=0, A0=0x7fff28fffb00, lda=-1807745072, sC=0x7ffeee6aaa00, sB=0x7ffe943ffe90)
    at /opt/conda/conda-bld/magma-cuda110_1611350834581/work/magmablas/spotf2_devicesfunc.cuh:323
#2  0x00005555a86de7c0 in _INTERNAL_61_tmpxft_00007a08_00000000_11_spotf2_kernels_compute_80_cpp1_ii_7b34eafc::spotf2_smlpout_anywidth_device (
    m=<optimized out>, n=0, A0=0x0, A=0x0, lda=<optimized out>, localstep=<optimized out>, gbstep=<optimized out>, info=<optimized out>)
    at /opt/conda/conda-bld/magma-cuda110_1611350834581/work/magmablas/spotf2_devicesfunc.cuh:450
#3  0x00005555a86fdb70 in spotf2_smlpout_anywidth_kernel_batched<<<(1,1,1),(2,8,1)>>> (m=2, n=2, dA_array=0x7ffeff7fffe0, ai=0, aj=0, lda=10,
    localstep=8, gbstep=0, info_array=0x7ffeff4ffff0, batchCount=4)
    at /opt/conda/conda-bld/magma-cuda110_1611350834581/work/magmablas/spotf2_kernels.cu:135
```
``` 
(cuda-gdb) info registers
R0             0x5c                92
R1             0xfffa40            16775744
R2             0x7fff              32767
R3             0x2                 2
R4             0x94400000          -1807745024
R5             0x7ffe              32766
```

Where an `LDG` is executed on `[R4]`.

Based on the `$pc` instructions, the most likely candidate is `A0` in:
```
static inline __device__ void sgemm_v20_1_anywidth_device(int m, int n, int k,
        const float* __restrict__ A0, int lda,
        float *sC, float  *sB)
{
[...]
        // prefetch next block. Azzam
        #ifdef ENABLE_COND5
        if (tx < m )  
        {      
        #endif
            #pragma unroll
            for (int i=0; i < POTF2_NB; i++)
            {
                rp[i] = A0[min(bound_A, tx + (i+(iter+POTF2_NB)) * lda)]; // min(bound,xxx) is to avoid reading out of bound
            }
        #ifdef ENABLE_COND5
        }
        #endif
```

I'm happy to dig further into it, if it makes sense.