@ngimel 2021-01-26
@ptrblck that's a very valid concern. Do you think one way to alleviate it would be to fill the padding element with 'nan' and run a battery of random tests? If that value is indeed read and processed, it would pollute the results. I don't want to leave the `nan` filling permanently, as that would be additional and unneded perf penalty, but I also don't think that existing tests in the test suite are enough to be reasonably sure that correct results are produced.