@ptrblck 2021-01-26
@ngimel Yes, I think it's a very good idea.
I can fill the `result` tensor with NaNs and run tests using multiples of 512 for the batch size first (as it was hitting the IMA) and then continue with random shapes in a reasonable amount of iterations (let me know, if you have specific shapes in mind).
If NaNs are detected in the result, I'll continue debugging it.